Update to Puzzle  - Changed calculation of continuous rows to be based on starting input values. Reverse continuous, continuous coloumn, etc are the same value

+++ Fifteen_Puzzle/MathUtil.h
@@ -17,7 +17,7 @@ public:
 	static uint64_t arrayHash(int* arrayPointer, int arraySize) {
 		uint64_t hashvalue = 0;
 		for (int i = 0; i < arraySize; i++) {
-			hashvalue += arrayPointer[i] * pow(16, i);
+			hashvalue += arrayPointer[i] * (int)pow(16, i);
 		}
 		return hashvalue;
 	}
